audio: move audio response functions to separate file
Move the play_tone_* and noop functions related to audio response to a
separate file. While at it, add a function init_audio_response to set
the function pointers.
