+++ main/audio.h
@@ -1,0 +1,1 @@
+void init_audio_response(void);
